changed u32 to usize, code more elegant

+++ head_color.rs
@@ -7,10 +7,10 @@ fn main() {
 
     let file_name = &env::args().nth(1).unwrap();
     let col_number_str = &env::args().nth(2).unwrap();
-    let col_number: u32 = col_number_str.parse::<u32>().unwrap();
+    let col_number: usize = col_number_str.parse::<usize>().unwrap();
     let delim = &env::args().nth(3).unwrap();
     let n_lines_str = &env::args().nth(4).unwrap();
-    let n_lines: u32 = n_lines_str.parse::<u32>().unwrap();
+    let n_lines: usize = n_lines_str.parse::<usize>().unwrap();
 
     let f = File::open(file_name).unwrap();
     let reader = BufReader::new(f);
@@ -29,11 +29,11 @@ fn main() {
         let colored_string = "".to_string() + green + my_vec[2] + nc;
 
         for i in 0..col_number {
-            print!("{}{}", my_vec[i as usize], delim);
+            print!("{}{}", my_vec[i], delim);
         }
         print!("{}", colored_string);
-        for i in ((col_number+1) as usize)..(my_vec.len()) {
-            print!("{}{}", delim, my_vec[i as usize]);
+        for i in ((col_number+1))..(my_vec.len()) {
+            print!("{}{}", delim, my_vec[i]);
         }
         println!("");
         row_number += 1;
